+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 ## 0.6.16
 * `holding` now relies on `HoldBuilder` for more natural composition of holds (#84).
+* Add a new way to create holding stores (#86).
+* Simple changes now all bear `isInitial` field (#85).
 
 ## 0.6.15
 * `holding` now automatically saves changes to objects implementing `Container` when those objects change (#79).